pyproject.toml
[project]
name = "profile-lg-lf"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.117.1",
    "httpx>=0.28.1",
    "langchain>=0.3.27",
    "langchain-openai>=0.3.33",
    "langfuse>=3.5.0",
    "langgraph>=0.6.7",
    "pydantic>=2.11.9",
    "pytest>=8.4.2",
    "uvicorn>=0.37.0",
    "python-dotenv>=1.0.0"
]
